requirements.txt
streamlit
streamlit-aggrid>=1.2.1.post2
altair<6

pandas
scipy
matplotlib>=3.10.8
scikit-learn>=1.8.0
seaborn>=0.13.2
umap-learn>=0.5.11
plotly>=6.6.0

python-dotenv
protobuf==3.20.*
pypdf

langchain
langgraph
langchain-core
langchain-openai
langchain-anthropic
langchain-community
langchain-model-profiles
langchain-tavily
langchain-chroma
langchain-huggingface
langchain-text-splitters
langchain-google-genai
langgraph-cli
langgraph-api
langsmith
tavily-python

google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib